requirements.txt
streamlit
Pillow
torch
torchvision
opencv-python
mediapipe
numpy
pandas
transformers
ultralytics
yolov5 @ git+https://github.com/ultralytics/yolov5.git
